backend/requirements.txt
annotated-types==0.6.0
anthropic==0.26.1
anyio==4.3.0
attrs==23.2.0
blis==0.7.11
boto3==1.34.111
botocore==1.34.111
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cohere==5.5.1
confection==0.1.4
cymem==2.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli
fastavro==1.9.4
filelock==3.14.0
fsspec==2024.5.0
groq==0.7.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.4
jiter==0.4.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistral_common==1.0.2
mistralai==0.1.8
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==24.0
preshed==3.0.9
psycopg2==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.18.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
safetensors==0.4.3
sentencepiece==0.1.99
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
sympy==1.12
textstat==0.7.3
thinc==8.2.3
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
types-requests==2.32.0.20240521
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
weasel
typer
websockets==12.0
